requirements.txt
tensorflow-cpu==2.4.1
requests==2.25.1
pillow==8.1.0
numpy==1.19.5
pyyaml==5.4.1
orjson==3.4.7
#uvicorn==0.13.3
#fastapi==0.63.0